Refactor CI
