+++ .github/workflows/main.yml
@@ -119,12 +119,12 @@ jobs:
       - uses: actions/checkout@v4
         with:
           submodules: recursive
-      - name: Set up Python ${{ matrix.python-version }}
+      - name: Set up Python
         uses: actions/setup-python@v5
         with:
-          python-version: ${{ matrix.python-version }}
+          python-version: "3.12"
       - name: Complete
         run: |
-          echo "Updating badges data..."
+          echo "Updating badges data for main workflow..."
           python3 .ci/update_badges.py main 5  # We test for 5 distinct Python versions
-          echo "CI completed"
+          echo "CI completed."
