+++ .github/workflows/nightly.yml
@@ -104,12 +104,12 @@ jobs:
       - uses: actions/checkout@v4
         with:
           submodules: recursive
-      - name: Set up Python ${{ matrix.python-version }}
+      - name: Set up Python
         uses: actions/setup-python@v5
         with:
-          python-version: ${{ matrix.python-version }}
+          python-version: "3.12"
       - name: Complete
         run: |
-          echo "Updating badges data for nightly..."
+          echo "Updating badges data for nightly workflow..."
           python3 .ci/update_badges.py nightly 5  # We test for 5 distinct Python versions
           echo "CI completed."
